+++ fugue-scala/src/test/scala/io/atlassian/fugue/law/package.scala
@@ -26,8 +26,8 @@ package object law {
     if (isEq.lhs().isInstanceOf[Comparable[_]] && isEq.lhs().asInstanceOf[Comparable[A]].compareTo(isEq.rhs()) == 0) proved
     else if (isEq.lhs().equals(isEq.rhs())) proved
     else falsified :| {
-      val exp = Pretty.pretty[A](isEq.lhs(), Pretty.Params(0))
-      val act = Pretty.pretty[A](isEq.rhs(), Pretty.Params(0))
+      val exp = Pretty.pretty[A](isEq.rhs(), Pretty.Params(0))
+      val act = Pretty.pretty[A](isEq.lhs(), Pretty.Params(0))
       "Expected " + exp + " but got " + act
     }
 }