+++ fugue/src/main/java/io/atlassian/fugue/Either.java
@@ -15,18 +15,19 @@
  */
 package io.atlassian.fugue;
 
-import java.io.Serializable;
-import java.util.Iterator;
-import java.util.NoSuchElementException;
-import java.util.function.Consumer;
-import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.function.Supplier;
-
 import static io.atlassian.fugue.Option.none;
 import static io.atlassian.fugue.Option.some;
 import static io.atlassian.fugue.Suppliers.ofInstance;
 import static java.util.Objects.requireNonNull;
+
+import java.io.Serializable;
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+import java.util.Optional;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 
 /**
  * A class that acts as a container for a value of one of two types. An Either
@@ -302,13 +303,43 @@ public abstract class Either<L, R> implements Serializable {
    * @param or Function to run if this is a left
    * @return contained value of R or result of {@code or}
    * @since 2.3
+   * @deprecated In favor of {@code rightOr}
+   * @see Either#rightOr(Function)
    */
-  public final R valueOr(final Function<L, ? extends R> or) {
+  @Deprecated public final R valueOr(final Function<L, ? extends R> or) {
+    return rightOr(or);
+  }
+
+  /**
+   * If this is a right return the contained value, else return the result of
+   * applying {@code leftTransformer} on left
+   *
+   * @param leftTransformer Function to run on left if this is a left
+   * @return contained value of right or result of {@code leftTransformer}
+   * @since 3.2
+   */
+  public final R rightOr(final Function<L, ? extends R> leftTransformer) {
     if (right().isDefined()) {
       return right().get();
     }
 
-    return or.apply(left().get());
+    return leftTransformer.apply(left().get());
+  }
+
+  /**
+   * If this is a left return the contained value, else return the result of
+   * applying {@code rightTransformer} on right
+   *
+   * @param rightTransformer Function to run on right if this is a right
+   * @return contained value of left or result of {@code rightTransformer}
+   * @since 3.2
+   */
+  public final L leftOr(final Function<R, ? extends L> rightTransformer) {
+    if (left().isDefined()) {
+      return left().get();
+    }
+
+    return rightTransformer.apply(right().get());
   }
 
   /**
@@ -324,6 +355,19 @@ public abstract class Either<L, R> implements Serializable {
    */
   public final Option<Either<L, R>> filter(final Predicate<? super R> p) {
     return right().filter(p);
+  }
+
+  /**
+   * Convert this Either to an {@link Optional}. Returns with
+   * {@link Optional#of(Object)} if it is a right, otherwise
+   * {@link Optional#empty()}.
+   *
+   * @return The right projection's value in <code>of</code> if it exists,
+   * otherwise <code>empty</code>.
+   * @since 4.0
+   */
+  public final Optional<R> toOptional() {
+    return right().toOptional();
   }
 
   /**
@@ -591,7 +635,11 @@ public abstract class Either<L, R> implements Serializable {
     }
 
     @Override public final Option<A> toOption() {
-      return isDefined() ? some(get()) : Option.<A> none();
+      return isDefined() ? some(get()) : none();
+    }
+
+    @Override public final Optional<A> toOptional() {
+      return toOption().toOptional();
     }
 
     @Override public final boolean exists(final Predicate<? super A> f) {
@@ -627,9 +675,7 @@ public abstract class Either<L, R> implements Serializable {
     }
 
     @Deprecated @Override public final void foreach(final Effect<? super A> f) {
-      if (isDefined()) {
-        f.apply(get());
-      }
+      this.forEach(f::apply);
     }
 
     @Override public final void forEach(final Consumer<? super A> f) {
@@ -709,7 +755,7 @@ public abstract class Either<L, R> implements Serializable {
      * Returns <code>None</code> if this projection has no value or if the given
      * predicate <code>p</code> does not hold for the value, otherwise, returns
      * a left in <code>Some</code>.
-     * 
+     *
      * @param <X> the RHS type
      * @param f The predicate function to test on this projection's value.
      * @return <code>None</code> if this projection has no value or if the given
@@ -827,14 +873,14 @@ public abstract class Either<L, R> implements Serializable {
      * @return An either after binding through this projection.
      */
     public <X> Either<L, X> sequence(final Either<L, X> e) {
-      return flatMap(Functions.<R, Either<L, X>> constant(e));
+      return flatMap(Functions.constant(e));
     }
 
     /**
      * Returns <code>None</code> if this projection has no value or if the given
      * predicate <code>p</code> does not hold for the value, otherwise, returns
      * a right in <code>Some</code>.
-     * 
+     *
      * @param <X> the LHS type
      * @param f The predicate function to test on this projection's value.
      * @return <code>None</code> if this projection has no value or if the given
@@ -904,6 +950,16 @@ public abstract class Either<L, R> implements Serializable {
      * otherwise <code>None</code>.
      */
     Option<? super A> toOption();
+
+    /**
+     * Returns this projection's value in {@link Optional#of} if it exists,
+     * otherwise {@link Optional#empty()}.
+     *
+     * @return This projection's value in <code>of</code> if it exists,
+     * otherwise <code>empty</code>.
+     * @since 4.0
+     */
+    Optional<? super A> toOptional();
 
     /**
      * The value of this projection or the result of the given function on the