+++ fugue/src/test/java/io/atlassian/fugue/EitherLeftBiasTest.java
@@ -1,0 +1,21 @@
+package io.atlassian.fugue;
+
+import static io.atlassian.fugue.Either.left;
+import static io.atlassian.fugue.Either.right;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+
+import org.junit.Test;
+
+public class EitherLeftBiasTest {
+  private final Either<String, Integer> l = left("heyaa!");
+  private final Either<String, Integer> r = right(12);
+
+  @Test public void leftOr_valueIsALeft_returnsLeftValue() {
+    assertThat(l.leftOr(Functions.constant("yo dude")), is("heyaa!"));
+  }
+
+  @Test public void leftOr_valueIsARight_returnsRightTransformerResult() {
+    assertThat(r.leftOr(Functions.constant("yo dude")), is("yo dude"));
+  }
+}