+++ fugue/src/test/java/io/atlassian/fugue/OptionSerializationTest.java
@@ -4,10 +4,12 @@ import org.junit.Test;
 
 import java.io.IOException;
 import java.io.NotSerializableException;
+import java.lang.reflect.Field;
+import java.util.Base64;
 
 import static io.atlassian.fugue.Serializer.toBytes;
 import static io.atlassian.fugue.Serializer.toObject;
-import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.assertThat;
 
 public class OptionSerializationTest {
@@ -23,5 +25,11 @@ public class OptionSerializationTest {
 
   @Test(expected = NotSerializableException.class) public void serializeSomeNonSerializable() throws IOException {
     toObject(toBytes(Option.some(Serializer.Unserializable.instance())));
+  }
+
+  @Test public void deserializeAnonymousNoneReadResolve() throws IOException {
+    final byte[] serialized = Base64.getDecoder().decode(
+      "rO0ABXNyABtpby5hdGxhc3NpYW4uZnVndWUuT3B0aW9uJDHki4wrMT+ZGgIAAHhyABlpby5hdGxhc3NpYW4uZnVndWUuT3B0aW9ubO2YatZzMVECAAB4cA==");
+    assertThat(toObject(serialized), sameInstance(Option.none()));
   }
 }